Changed wavelength region calc'ed by fluspect
Edited the fluo_param.f90 file to make sure the size of the arrays it
was allocating corresponded to the optipar fluspect files. This means
that only wavelengths from 400 to 2400 nm are calculated. We believe
there was a bug in the way this was done, as extra values (including
a zero) from the wrong fluspect coefficients column were being read in
by the opticoef commands.