Biochemical and Fluo eq'ns proceed without error
The model now proceeds through the fluorescence and biochemical
equations without error. However, the GPP calculated is not being
accepted into the cbalance subroutine. Therefore the cbalance call is
commented out here. Also, the SCOPE GPP (zgppfluo) is not being filled
with the fluo.f90 GPP data (Agtot). Need to add this in to ensure it is
pulling each grid fraction calculated Agtoto and putting it into the
vegetation point array zgppfluo.