Code set to run for 1day(1month) at diff. timestep
Changed the timestep at which we run the fluorescence routine to be
12 hours different, at 12 instead of 24. We need to clear up issues with
how the time steps are handled, whether it actually is done by "hours
until next solar zenith", or by actual 24-hours time. Also here, the
bethy.g90 is set to run for just one inner daily time step so we could
print out the scope GPP and Fluo at all vegetation points (all vp's are
run here, instead of 1 in 100). We see that reasonable values are
output.