Determined error with monthly output, temp fix
The output of SCOPE fluo and gpp was unreasonably high in Jan, May and
Oct (months 1,5,10). This was because of issues with dayint=25, where
it would run scope twice in those months (every 25th doy). For now we
have fixed this by changing dayint in the control file to 32. Longer
term fix is needed. Issues now arise with whether the model is actually
getting a representative day for each month, instead just picking out
the climate data for what ever real-day (rday) it is.