Added comments to display time-structure
Removed the .bin write out of SCOPE gpp and fluo. Added in comments to
show the time-structure of the model code from bethy.f90 downwards when
running. This is to help understand what to do with parallelizing the
code: found that there are multiple vp/gridp loops (e.g. in photo1,
photo2, fluo) making parallelizing at bethy.f90 more difficult.