Moving modtran file reading outside of vp loop
In the middle of moving all the modtran file (summer, winter, tropics)
actions that occur within the vp loop outside the loop. This has
required re-writing of a couple of subroutines. Also we had to move
the reading of the files outside of fluo.f90 and into fluo_param.f90.