Added psi, tto, nli to OMP SHARED list
Becuase these variables were not in the SHARED list, errors were
propagating absurdly large numbers through code leading to errors.
Also added USE ONLY statements to chemical and rtmf. I think it is set
up in a way that all the variables are set up in fluo_param, so the
USE ONLY statements in fluorescence only requires the subroutine names