Added kClrel fluspect var as output, del .dat out
Commented out the savefield calls in mo_io that write .dat output.
Added in another fluspect output variable to the argument list, also
removed it as a module variable and made it fluspect-subroutine
specific, so its declared in each routine that uses it. This is to
prevent overwriting during parallelization.