Threadprivate for fluspects modules vars
Determined what variables in fluspect are defined as module variables,
but given a value (on LHS of =) in fluspect (did this on a debug
branch). Then set these in a OMP THREADPRIVATE directive in fluo_param.
For "tran" and "refl" they weren't allocated by the time we got into
fluspect so I had to put in a temp fix using IF NOT ALLOCATED
statements.