Start altering leafangles, called in vp loop
We want leafangles (leaf angle distribution function) call to be done
within the vp loop so that we can assign PFT specific (from parameter
input vp array) values for leaf angle parameters (LIDFa, LIDFb).
Here we have started this, by adding in the call, put the subroutine
into the USE statement, and added the arrays to the shared OMP list.