Added shortwave down forcing scaling (Tol Kn)
Added in some lines of code to the mo_climate.f90 module where
shortwave-down radiation forcing is read in. There are three lines
that essentially allow us to "scale" the swdown forcing by any amount.
The scaling factor could be called a parameter now, and there is one
for each latitudinal band (Sth Hemis: -90 to -30,Tropics: -30 to 30,
Nth Hemis: 30 to 90).