Restore to original standard jacobian calc runs
We have returned chemical (Kn model) and mo_climate (swdown scaling)
to the original form as was used in the calculation of standard
jacobians.

We also made some adjustments to the timings used in SCOPE. This change
should allow diurnal runs to go ahead (however the output still needs
to be adjusted to save diurnal time steps separately).