Bug fix: A solar zenith angle calc in SCOPE fixed
There was an error in how SCOPE solar zenith angle calculation was done
in the SUBROUTINE calczenithangle. This error was in how it calc'd time
when sun was at it's highest angle (variable "tm"). This calc took into
account longitude, when it shouldn't be as we're running grid points on
local time. This meant that even in Jan at midday, some regions such as
Australia had solar zenith angle >90 degrees. This fix also improved
the problem of high SIF over central Australia.