Bug fix: DOY calc in SCOPE fixed
SCOPE calculated the DOY (used in the solar zenith angle calculation)
to be about 1 month off what it should have been. We have reconciled
this by using the bethy time variables iday0 and iday1 to figure out
the mid-point DOY needed (mid-point of the forcings averaging period).
iday0 and iday1 are now argument inputs to the fluorescence call.