Modified how climate forcing is averaged for photo
The photosynthesis modules (including BETHY phsyn1/2 and SCOPE) are
only run every "dayint", while hydrology and phenology are run daily.
The climate forcing for the photo modules is determined by averaging
between dayint, but this can result in overlap of months forcing. So
for our purposes when running just one day per month, we determine the
month for which the timestep is at, then calculate the actual monthly
mean climate conditions using an iday0 and iday1 corresponding to the
start and end of the month, independent of dayint.