vp block split parallelisation set-up
Set-up the code so that it can be run on NCI and split into separate
jobs based upon "blocks" of vegetation points. This set-up now requires
two extra variables set in the control namelist file (nblocks,iblock)
that tell the code how many blocks to split it in to (i.e.
jobs=nblocks). Each job will have it's own control file, with a
different outdir and iblock set.