Running on NCI and matches Vortex runs
Updated the files, input paths and debugged so that it can now run on
NCI (Raijin). To note in particular: There were problems with OpenMP
here on raijin, such that thread-private arrays that were allocated
outside of the parallel block lost their allocation once in the
parallel block of code. To handle this issue, we made the default mode
as firstprivate, then placed private statements for an local arrays
that are actually allocated within the parallel block. Also, no
vp-block split parallelisation in this commit. That still needs work.