Standard setup for Jacobian runs (block parallel)
This version was used to construct the jacobian (sensitivity)
experiments and the standard model run for 2010, with full diurnal
cycle of GPP and SIF. This included steps to block-parallelize the code
where the model is split into blocks for separate jobs. To run this
block parallelized version, use run_post_block and set the name of the
experiment in the output/merge_outputs.sh file that stitches together
the separate job output .nc files.