Solved iss02: scope output in daily nc files
rfluo and rgppfluo are now written out in netcdf files that are of the
size ndays-simulated x 24. This means there is no averaging per month
required when writing the output, as there previously was.

We can still calculate monthly mean climate forcing to drive photo
calculations by commenting out some iday0,iday1 variables in bethy and
uncommenting some other lines lower down that determine the first and
last day of the month.