Setup for hires runs for SCOPE monthly mean
This commit is now setup to run scope sif and gpp at 1pm driven by
monthly mean climate forcing.

There is an additional file committed which is a python script used
to concatenate the output files from the block parallelisation runs
into a single netcdf file for scope sif and gpp. This .py script runs
in conjunction with the merge_outputs.sh script.