Set config to optionally run selected gp's
Have setup mo_config to run only selected grid-points and their
associated veg-points. You can now specify a variable in the control
file called blockvpfile, that tells you what veg-points to run, which
can also be split up over blocks for block parallelisation. This is
compatible with the previous block parallelisation as well (just don't
specify a blockvpfile and it will split blocks as previously). To run
all veg-points in a block-parallel way, we recommend using the
blockvpfile and the create_blockvp.py script as it splits cleanly
between grid-points.