Setup option for prescribing LAI to SCOPE fluo
You can now set a lai text file to feed into the model to prescribe
lai to scope fluorescence calculations. The prescribed lai is read in
the mo_climate module, and fed to fluo via bethy. To switch it off,
just comment out that line in the bethy module.
In doing this, we realized that if the lai given to fluo is zero, then
there is an double allocation error, as some deallocate statements were
within the lai if statement. We have remedied this.