Fixed error in allocating arrays in fluo
When switching from prescribed lai back to bethy prognostic lai for
scope, errors arose. Moving the allocate statements for a few of the
arrays in the fluo vp loop (inside the zlai if test) to outside the
zlai if test fixed the issue.
Running this version of bethy-scope yields the same results in sif as
the standard run (par4000) in the oco2 jacobian calculations.