Timing was wrong for BETHY monthly mean calcs
When running BETHY on the same monthly mean forcing as SCOPE is run,
we alter the iday0, iday1 and dayint variables, but when writing out
BETHY fields (e.g. rgpp) it requires dayint. There is a line in the
bethy subroutine that alters dayint if the interval extends over the
total run period (tdays), this can be tripped when we choose monthly
mean forcing. So we should instead use the original iday0,iday1,dayint
code, and choose a lower dayint (e.g. 10) when running BETHY stand-
alone, and let diagnostics split the data accordingly.