SCOPE v1.60 update: Ci and effcon
Updated the calculation of Ci to be different for C3 and C4 plants.
This is slightly different to SCOPE v1.60 though as I believe there
is a bug in their calculation of Ci as it does not use the constant
value of 1.6 (see Franks et al., 2017, Eq. 9). We also updated the
calculation of effcon as per SCOPE v1.60.