Explicity calculation of spfy
The relative specificity of Rubisco (spfy) was set as a constant with
a Q10 temperature function. However, we have altered this to calculate
it explicitly. We did this because the ratio of Vomax/Vcmax (avovc)
is a highly uncertain kinetic parameter (see Von Caemmerer, 2000) and
spfy is sensitive to this parameter. This is different to SCOPE
stand-alone.