Update to get openmp working with fluspect update
Had to make a couple of changes. i) fixed a floating invalid in
mo_rtmo. ii) had to alter the run script (will add that in next
commit) for pbsdsh. NOTE: there is still an error occurring when I run
only with block parallelization. Not sure where this is coming from.