Added feature to read site LAI
Can now read a site LAI file. The file name is given in the control
file by "site_file_lai". If there is no "site_file_lai", then there
will be no LAI read into the "dlai" array. The dlai array is a
(vp, time in days) dimension array. This means the site_file_lai must
be given as (vp, time in days); e.g. if it is a 1 year simulation with
one vp, then it should be one line with 365 values (f8.2).