Fixed error in wlF calculations
An error was present in the fluo_param module that allowed for two
different ways of calculating the wavelengths of fluorescence (wlF)
and the frequency (wavelength) chosen to write rfluo output at
(ifreq_sat). This bug did not cause errors in how fluorescence was
calculated or written out, but it can cause an error when we re-define
the wavelength range for fluorescence. This new setup allows for
any range of fluorescence wavelengths to be simulated using the
variables: freq_Fsmax, freq_Fsmin, freq_sat.