Testing vertical profile calc of yields
Testing the calculation and writing out of the relative fluorescence
quantum yield parameter (eta) for sunlit (etau) and shaded (etah)
leaves. This commit working and writes out values, but it is
calculating the integrated eta over angles and layers, as the 'angles'
only option is not working yet. I.e. this is not top-of-canopy yield.